# openapi.yaml
openapi: 3.0.3
info:
  title: Formance Simple orchestration Service API
  version: 0.1.0
paths:
...
  /triggers:
    get:
      summary: List triggers
      operationId: listTriggers
      description: List triggers
      tags:
        - orchestration.v1
      parameters:
        - name: name
          in: query
          description: search by name
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: List of triggers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTriggersResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
      security:
        - Authorization:
            - orchestration:read
    post:
      summary: Create trigger
      operationId: createTrigger
      description: Create trigger
      tags:
        - orchestration.v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerData'
      responses:
        '201':
          description: Created trigger
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTriggerResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
      security:
        - Authorization:
            - orchestration:write
...
  /workflows:
    get:
      summary: List registered workflows
      operationId: listWorkflows
      description: List registered workflows
      tags:
        - orchestration.v1
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: List of workflows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListWorkflowsResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
      security:
        - Authorization:
            - orchestration:read
    post:
      summary: Create workflow
      operationId: createWorkflow
      description: Create a workflow
      tags:
        - orchestration.v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
      responses:
        '201':
          description: Created workflow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateWorkflowResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
      security:
        - Authorization:
            - orchestration:write
...
components:
  schemas:
...
    Error:
      type: object
      required:
        - errorCode
        - errorMessage
      properties:
        errorCode:
          type: string
          enum:
            - VALIDATION
            - NOT_FOUND
            - INTERNAL
        errorMessage:
          type: string
    WorkflowConfig:
      type: object
      required:
        - stages
      properties:
        name:
          type: string
        stages:
          type: array
          items:
            type: object
            additionalProperties: {}
    Workflow:
      type: object
      required:
        - config
        - createdAt
        - updatedAt
        - id
      properties:
        config:
          $ref: '#/components/schemas/WorkflowConfig'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        id:
          type: string
...
    ListWorkflowsResponse:
      type: object
      properties:
        data:
          items:
            $ref: '#/components/schemas/Workflow'
          type: array
        pageSize:
          type: integer
        hasMore:
          type: boolean
        previous:
          type: string
        next:
          type: string
      required:
        - data
        - pageSize
        - hasMore
    TriggerData:
      type: object
      required:
        - event
        - workflowID
      properties:
        event:
          type: string
        workflowID:
          type: string
        version:
          type: string
        filter:
          type: string
        vars:
          type: object
          additionalProperties: true
        name:
          type: string
    Trigger:
      allOf:
        - $ref: '#/components/schemas/TriggerData'
        - type: object
          required:
            - id
            - createdAt
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
...
    ListTriggersResponse:
      type: object
      properties:
        data:
          items:
            $ref: '#/components/schemas/Trigger'
          type: array
        pageSize:
          type: integer
        hasMore:
          type: boolean
        previous:
          type: string
        next:
          type: string
      required:
        - data
        - pageSize
        - hasMore
    CreateWorkflowRequest:
      $ref: '#/components/schemas/WorkflowConfig'
    CreateWorkflowResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/Workflow'
    CreateTriggerResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/Trigger'
...
  parameters:
    PageSize:
      name: pageSize
      in: query
      description: |
        The maximum number of results to return per page.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 15
        format: int64
      example: 100
    Cursor:
      name: cursor
      in: query
      description: |
        Parameter used in pagination requests.
        Set to the value of next for the next page of results.
        Set to the value of previous for the previous page of results.
        No other parameters can be set when this parameter is set.
      required: false
      schema:
        type: string
        example: aHR0cHM6Ly9nLnBhZ2UvTmVrby1SYW1lbj9zaGFyZQ==
  responses:
    ErrorResponse:
      description: General error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    Authorization:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: "/oauth/token"
          refreshUrl: "/oauth/token"
          scopes: {}
servers:
  - url: http://localhost:8080/